+++ exporter/exporterhelper/internal/persistent_storage_test.go
@@ -57,7 +57,7 @@ func createTestPersistentStorage(client storage.Client) *persistentContiguousSto
 type fakeTracesRequest struct {
 	td                         ptrace.Traces
 	processingFinishedCallback func()
-	PersistentRequest
+	Request
 }
 
 func newFakeTracesRequest(td ptrace.Traces) *fakeTracesRequest {
@@ -81,7 +81,7 @@ func (fd *fakeTracesRequest) SetOnProcessingFinished(callback func()) {
 }
 
 func newFakeTracesRequestUnmarshalerFunc() RequestUnmarshaler {
-	return func(bytes []byte) (PersistentRequest, error) {
+	return func(bytes []byte) (Request, error) {
 		traces, err := ptrace.NewProtoUnmarshaler().UnmarshalTraces(bytes)
 		if err != nil {
 			return nil, err
@@ -416,8 +416,8 @@ func TestPersistentStorage_ItemIndexMarshaling(t *testing.T) {
 	}
 }
 
-func getItemFromChannel(t *testing.T, pcs *persistentContiguousStorage) PersistentRequest {
-	var readReq PersistentRequest
+func getItemFromChannel(t *testing.T, pcs *persistentContiguousStorage) Request {
+	var readReq Request
 	require.Eventually(t, func() bool {
 		readReq = <-pcs.get()
 		return true