+++ pdata/internal/common.go
@@ -371,16 +371,8 @@ func (v Value) Equal(av Value) bool {
 			return false
 		}
 
-		for i, val := range avv {
-			val := val
-			newAv := newValue(&vv[i])
-
-			// According to the specification, array values must be scalar.
-			if avType := newAv.Type(); avType == ValueTypeSlice || avType == ValueTypeMap {
-				return false
-			}
-
-			if !newAv.Equal(newValue(&val)) {
+		for i := range avv {
+			if !newValue(&vv[i]).Equal(newValue(&avv[i])) {
 				return false
 			}
 		}
@@ -394,13 +386,13 @@ func (v Value) Equal(av Value) bool {
 
 		m := newMap(&avv)
 
-		for _, val := range cc {
-			newAv, ok := m.Get(val.Key)
+		for i := range cc {
+			newAv, ok := m.Get(cc[i].Key)
 			if !ok {
 				return false
 			}
 
-			if !newAv.Equal(newValue(&val.Value)) {
+			if !newAv.Equal(newValue(&cc[i].Value)) {
 				return false
 			}
 		}
@@ -452,7 +444,7 @@ func (v Value) AsString() string {
 // This allows us to avoid using reflection.
 func float64AsString(f float64) string {
 	if math.IsInf(f, 0) || math.IsNaN(f) {
-		return fmt.Sprintf("json: unsupported value: %s", strconv.FormatFloat(f, 'g', -1, int(64)))
+		return fmt.Sprintf("json: unsupported value: %s", strconv.FormatFloat(f, 'g', -1, 64))
 	}
 
 	// Convert as if by ES6 number to string conversion.
@@ -467,7 +459,7 @@ func float64AsString(f float64) string {
 	if abs != 0 && (abs < 1e-6 || abs >= 1e21) {
 		fmt = 'e'
 	}
-	b = strconv.AppendFloat(b, f, fmt, -1, int(64))
+	b = strconv.AppendFloat(b, f, fmt, -1, 64)
 	if fmt == 'e' {
 		// clean up e-09 to e-9
 		n := len(b)